Tests and bugfixes that some of those tests revealed.
There is a lot of annoying code duplication in the tests, and I tried
a macro that generates the inner functions within the test functions,
something like

    #[test]
    fn test_something() {
        check_autodiff!(
            T::one() + x,
            1f64,
            -5..5
        )
    }

would generate

    #[test]
    fn test_something() {
        fn f<T>(x: T) -> T where T: From<f64> + Float {
            T::one() + x
        }

        fn df_dx(x: f64) -> f64 {
            1f64
        }

        for val in -5..5 {
            let x = val as f64;
            let adval = f(AutoDiff::var(x));
            let fval = f(x);
            let dval = df_dx(x);

            assert_almost_equal!(adval.val(), fval);
            assert_almost_equal!(adval.dval(), dval);
        }
    }

But it doesn't seem to work, and the macro invocation requires the name
x to be in scope, when matching it as an ident in the macro def. Don't
know how to get around atm.